move github env definition
