+++ sources/src/actions/github-env.ts
@@ -1,11 +1,11 @@
-import {GradleContext, GradleEnvImplementation, GradleEnvStateImplementation, GradleEnv} from './env'
+import {GradleContext, GradleEnvImplementation, GradleEnvStateImplementation, GradleEnv} from '../env/env'
 import * as core from '@actions/core'
 import * as cache from '@actions/cache'
 import * as exec from '@actions/exec'
 import * as github from '@actions/github'
 import * as glob from '@actions/glob'
-import {CacheEntryAlreadyExistsError, CacheValidationError, GradleEnvCache, GradleEnvCacheEntry} from './cache'
-import {GradleGlob} from './glob'
+import {CacheEntryAlreadyExistsError, CacheValidationError, GradleEnvCache, GradleEnvCacheEntry} from '../env/cache'
+import {GradleGlob} from '../env/glob'
 
 const githubContext: GradleContext = {
     workflowIdentifier: github.context.workflow,
