+++ sources/src/actions/setup-gradle/post.ts
@@ -4,7 +4,7 @@ import {handlePostActionError} from '../../errors'
 import {emitDeprecationWarnings, restoreDeprecationState} from '../../deprecation-collector'
 import {SetupGradleAction} from '../../setup-gradle'
 import {setupDependencies} from '../../inject'
-import {githubActionGradleEnv} from '../../env/github-action'
+import {githubActionGradleEnv} from '../github-env'
 
 // Catch and log any unhandled exceptions.  These exceptions can leak out of the uploadChunk method in
 // @actions/toolkit when a failed upload closes the file descriptor causing any in-process reads to
